+++ RELEASE_NOTES.md
@@ -1,3 +1,22 @@
+# 0.2.96
+
+Changes:
+
+ * fixed RunScanner overwriting user-marked run completion
+
+# 0.2.95
+
+Changes:
+
+ * Show average insert sizes on List Pools page
+ * Add support for NovaSeq
+ * Fixed missing preMigrationId after bulk saving Libraries
+ * Fixed saving boxes with blank barcode
+ * Fixed user data (roles) not updating from LDAP
+ * Added default BoxUse and BoxSize
+
+Bad: RunScanner could sometimes overwrite user-marked run completion
+
 # 0.2.94
 
 Changes:
@@ -5,6 +24,8 @@ Changes:
  * Fix project overview dialog
  * Add more numbers for pagination
  * Add new Brady THT-179-492 labels
+
+Bad: RunScanner could sometimes overwrite user-marked run completion
 
 # 0.2.93